Remove buildSierraBibs from build-es-document
Remove buildSierraBibs because it wasn't really serving its name any
more given that all records are wrapped immediately in Sierra model
classes. The one functional purpose it served was to add back-references
to the parent SierraBib of SierraItems and SierraHoldings. This function
has been re-sited in the modelPrefetch function, since that's where
we're currently building most of that hierarchy.