Ensure contributor* fields don't duplicate creator*
Updates contributorLiteralNormalized and contributorLiteralWithoutDates fields
to remove entries that are found to duplicate creator* fields.

This is to ensure we don’t ever have a bib with, for example, creatorLiteralNormalized "Charles Dickens" and contributorLiteralNormalized "Charles Dickens" due to the two different Charles's being named in different roles, which would unfairly boost this record when matching on both of those Normalized fields.

Also adds removal of trailing period characters in creator and contributor fields so that we don't ever index e.g. "Dickens, Charles.".

https://newyorkpubliclibrary.atlassian.net/browse/SCC-4275